+++ common/src/main/java/net/more_rpg_classes/custom/SpellBuilderHelper.java
@@ -24,6 +24,22 @@ public class SpellBuilderHelper {
         spell.tooltip.show_activation = false;
         return spell;
     }
+    public static Spell.TargetCondition healthRangeCondition(float below, float above) {
+        var deadCondition = new Spell.TargetCondition();
+        deadCondition.health_percent_below = below;
+        deadCondition.health_percent_above = above;
+        return deadCondition;
+    }
+
     /// COLORS
     public static final Color ORANGE_COLOR = new Color(255.0F, 165.0F, 0.0F);
+    public static final Color GOLD = Color.from(0xffd700);
+    public static final Color CYAN = Color.from(0x00ffff);
+    public static final Color BRIGHT_GREEN = Color.from(0x8efea1);
+    public static final Color BRIGHT_CYAN = Color.from(0x83f4e6);
+    public static final Color MID_MAGENTA = Color.from(0xa082ce);
+    public static final Color MAGENTA = Color.from(0x7c37f5);
+    public static final Color BRIGHT_MAGENTA = Color.from(0xdd6af7);
+    public static final Color YELLOW_GREEN = Color.from(0xaffe0d);
+
 }